+++ packages/davinci-prisma/src/decorators.ts
@@ -1,6 +1,10 @@
 import { decorate, DecoratorId } from '@davinci/reflector'
-import { FullTextIndex, Index, UniqueIndex } from '@pmaltese/prisma-schema-generator'
-import { FieldDecoratorMeta, FieldDecoratorOptions } from './types'
+import {
+  FieldDecoratorMeta,
+  FieldDecoratorOptions,
+  ModelDecoratorMeta,
+  ViewDecoratorMeta,
+} from './types'
 
 export const field = (data?: FieldDecoratorOptions) =>
   decorate(<FieldDecoratorMeta>{ [DecoratorId]: 'davinci.prisma.field', data }, ['Property'], {
@@ -8,19 +12,13 @@ export const field = (data?: FieldDecoratorOptions) =>
     inherit: true,
   })
 
-export const model = (data?: {
-  name?: string
-  map?: string
-  indexes?: Array<Index>
-  uniqueIndexes?: Array<UniqueIndex>
-  fullTextIndexes?: Array<FullTextIndex>
-}) =>
+export const model = (data?: ModelDecoratorMeta['data']) =>
   decorate({ [DecoratorId]: 'davinci.prisma.model', data }, ['Class'], {
     allowMultiple: true,
     inherit: true,
   })
 
-export const view = (data?: { name?: string; map?: string }) =>
+export const view = (data?: ViewDecoratorMeta['data']) =>
   decorate({ [DecoratorId]: 'davinci.prisma.view', data }, ['Class'], {
     allowMultiple: true,
     inherit: true,